requirements.txt
asgiref==3.7.2
attrs==23.1.0
dj-database-url==2.0.0
Django==4.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.3
inflection==0.5.1
jsonschema==4.17.3
psycopg2-binary==2.9.6
PyJWT==2.7.0
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
